backend/requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
supabase>=2.4.0
anthropic>=0.28.0
python-dotenv>=1.0.1
httpx>=0.27.0
apscheduler>=3.10.4
pydantic>=2.7.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.9
requests>=2.31.0
